requirements.txt
alembic==1.7.7
cassandra-driver==3.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
databricks-cli==0.16.4
docker==5.0.3
entrypoints==0.4
Flask==2.0.3
Flask-Cors==3.0.10
geomet==0.2.1.post1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.0
joblib==1.1.0
Mako==1.2.0
MarkupSafe==2.1.1
mlflow==1.24.0
numpy==1.21.5
packaging==21.3
pandas==1.1.5
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
protobuf==3.19.4
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pywin32==227
PyYAML==6.0
querystring-parser==1.2.4
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
smmap==5.0.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
tabulate==0.8.9
threadpoolctl==3.1.0
tqdm==4.63.1
typing-extensions==4.1.1
urllib3==1.26.9
waitress==2.1.1
websocket-client==1.3.2
Werkzeug==2.0.3
zipp==3.7.0
